+++ include/sysdep.h
@@ -366,13 +366,13 @@ typedef unsigned long       uintptr_t;
 #    define MYFLT2LRND(x) (x > LONG_MIN && x < (double)LONG_MAX ? \
                            (int32) lrintf((float) (x)) : 0)
 #  else
-#    define MYFLT2LONG(x) (x > LONG_MIN && x < LONG_MAX ? \
+#    define MYFLT2LONG(x) (x > LONG_MIN && x < (double)LONG_MAX ? \
                            (int32) lrint((double) (x)) : 0)
-#    define MYFLT2LRND(x) (x > LONG_MIN && x < LONG_MAX ? \
+#    define MYFLT2LRND(x) (x > LONG_MIN && x < (double)LONG_MAX ? \
                            (int32) lrint((double) (x)) : 0)
-#    define MYFLT2LONG64(x) (x > LONG_MIN && x < LONG_MAX ? \
+#    define MYFLT2LONG64(x) (x > LONG_MIN && x < (double)LONG_MAX ? \
                            (int64_t) lrintl((double) (x)) : 0)
-#    define MYFLT2LRND64(x) (x > LONG_MIN && x < LONG_MAX ? \
+#    define MYFLT2LRND64(x) (x > LONG_MIN && x < (double)LONG_MAX ? \
                            (int64_t) lrintl((double) (x)) : 0)
 #  endif
 #elif defined(MSVC)
